requirements.txt
numpy
Pillow
argparse 
